+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-Features:
-- Add zone management to the API.
-- Add the request to get account balance to the account api.
+Fix:
+- Set CreateZoneRequest to return the instance of the request on the setters.